Preparation for OrderedRDD2.
Use off-heap storage.
Replaced Platform => is.hail.annotations.Memory.
Offsets are now Long, including stored offests to binary and arrays.

Add finalizer to MemoryBuffer.

+++ src/main/scala/is/hail/annotations/ReadRowsRDD.scala
@@ -284,7 +284,7 @@ class RichRDDRow(val rdd: RDD[Row]) extends AnyVal {
           rvb.addRow(t, r)
           val offset = rvb.end()
 
-          val rowSize = region.offset
+          val rowSize = region.offset.toInt
           out.writeInt(rowSize)
 
           en.writeRegionValue(f, region, offset)
